requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
supabase
groq
python-dotenv
pytest
